Move LogMessage unit test to vespalog and cleanup the test code.
